New Feature: Support for alternative configuration files
You are able to „overwrite“ the default invade.yml.dist file with your
own.

Just create a „invade.yml.dist“ - file and place it at:
„~/.invade/invade.yml.dist“.

You also able to create a project related default file by placing a
„invade.yml.dist“ at the projects root directory. It could be shared
with other users of a project by committing it to the repository.